+++ src/falconpy/__init__.py
@@ -94,6 +94,7 @@ from .aspm import ASPM
 from .certificate_based_exclusions import CertificateBasedExclusions
 from .cloud_aws_registration import CloudAWSRegistration
 from .cloud_azure_registration import CloudAzureRegistration
+from .cloud_oci_registration import CloudOCIRegistration
 from .cloud_snapshots import CloudSnapshots
 from .compliance_assessments import ComplianceAssessments
 from .configuration_assessment_evaluation_logic import ConfigurationAssessmentEvaluationLogic
@@ -215,7 +216,8 @@ __all__ = [
     "ImageAssessmentPolicies", "APIIntegrations", "ThreatGraph", "ExposureManagement",
     "CertificateBasedExclusions", "ComplianceAssessments", "HostMigration", "QuickScanPro",
     "DataScanner", "SensorUsage", "Downloads", "DeliverySettings", "ASPM", "IntelligenceFeeds",
-    "NGSIEM", "CorrelationRules", "CloudAWSRegistration", "CloudAzureRegistration"
+    "NGSIEM", "CorrelationRules", "CloudAWSRegistration", "CloudAzureRegistration",
+    "CloudOCIRegistration"
     ]
 """
 This is free and unencumbered software released into the public domain.